+++ _codegen/compat_fabric.py
@@ -16,6 +16,10 @@ import compat_core
 def _vt(ver):
     return tuple(int(x) for x in ver.split("-")[0].split("."))
 
+
+def is_26(ver):
+    return _vt(ver) >= (26, 0, 0)
+
 # ---- boat package move at MC 1.21.11 (vehicle.* -> vehicle.boat.*) ----
 # Used by a one-line cog conditional on each affected import; simple names are unchanged.
 def boat_pkg(ver):
@@ -130,7 +134,7 @@ def emit_client_imports(cog, loader, ver):
         cog.outl("import com.kishku7.lavaboats.client.LavaBoatRenderer;")
     else:
         cog.outl("import com.kishku7.lavaboats.client.LavaBoatLayers;")
-        cog.outl("import net.fabricmc.fabric.api.client.rendering.v1.EntityModelLayerRegistry;")
+        cog.outl("import net.fabricmc.fabric.api.client.rendering.v1." + _model_layer_registrar(ver) + ";")
         cog.outl("import net.minecraft.client.renderer.entity.BoatRenderer;")
         cog.outl("import " + boat_model_pkg(ver) + ".BoatModel;")
 
@@ -146,9 +150,10 @@ def emit_client_body(cog, loader, ver):
             cog.outl("{0}.register(ModEntities.{1},".format(reg, field))
             cog.outl("        ctx -> new LavaBoatRenderer(ctx, LavaBoatRenderer.{0}, {1}));".format(tex, "true" if chest else "false"))
     else:
+        mlr = _model_layer_registrar(ver)
         for field, tex, chest in _BOATS:
             maker = "createChestBoatModel" if chest else "createBoatModel"
-            cog.outl("EntityModelLayerRegistry.registerModelLayer(LavaBoatLayers.{0}, BoatModel::{1});".format(field, maker))
+            cog.outl("{0}.registerModelLayer(LavaBoatLayers.{1}, BoatModel::{2});".format(mlr, field, maker))
         for field, tex, chest in _BOATS:
             cog.outl("{0}.register(ModEntities.{1}, ctx -> new BoatRenderer(ctx, LavaBoatLayers.{1}));".format(reg, field))
 
@@ -163,5 +168,34 @@ def emit_join_unlock(cog, ver):
         arg = "LavaBoats.RECIPES.toArray(new ResourceLocation[0])"
     else:
         arg = "LavaBoats.RECIPES"
+    player = "handler.player" if is_26(ver) else "handler.getPlayer()"
     cog.outl("ServerPlayConnectionEvents.JOIN.register((handler, sender, server) ->")
-    cog.outl("        handler.getPlayer().awardRecipesByKey({0}));".format(arg))
+    cog.outl("        {0}.awardRecipesByKey({1}));".format(player, arg))
+
+# ---- fabric model-layer registry renamed EntityModelLayerRegistry -> ModelLayerRegistry at MC 26 ----
+def _model_layer_registrar(ver):
+    return "ModelLayerRegistry" if is_26(ver) else "EntityModelLayerRegistry"
+
+
+# ---- creative-tab insertion: ONE facade point (ModItems.init), version-branched. pre-26 uses
+#      itemgroup.v1 ItemGroupEvents; MC 26 uses creativetab.v1 CreativeModeTabEvents (fapi renamed it). ----
+def emit_moditems_tab_imports(cog, ver):
+    cog.outl("import net.minecraft.world.item.CreativeModeTabs;")
+    if is_26(ver):
+        cog.outl("import net.fabricmc.fabric.api.creativetab.v1.CreativeModeTabEvents;")
+    else:
+        cog.outl("import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;")
+
+
+def emit_moditems_tab(cog, ver):
+    cog.outl("// Slot the boats into the Tools & Utilities tab alongside the vanilla boats.")
+    if is_26(ver):
+        cog.outl("CreativeModeTabEvents.modifyOutputEvent(CreativeModeTabs.TOOLS_AND_UTILITIES).register(output -> {")
+        cog.outl("    output.accept(CRIMSON_BOAT);")
+        cog.outl("    output.accept(CRIMSON_CHEST_BOAT);")
+        cog.outl("    output.accept(WARPED_BOAT);")
+        cog.outl("    output.accept(WARPED_CHEST_BOAT);")
+        cog.outl("});")
+    else:
+        cog.outl("ItemGroupEvents.modifyEntriesEvent(CreativeModeTabs.TOOLS_AND_UTILITIES).register(entries ->")
+        cog.outl("        TAB_ITEMS.forEach(entries::accept));")